Cleaner abstraction for common keyword patterns
"Keywords" are a way that host HTMLBars environments shim into the
very lowest levels of the HTMLBars rendering system without directly
overriding the implementation of `inline`, `block`, etc.

Essentially, they give you an opportunity to examine the state of the
rendering system when a statement is encountered, and optionally mutate
that state, proceed as usual, or call an entirely different hook.

As an example, the `{{outlet}}` hook in Ember is implemented as a
keyword. Despite being an inline statement (i.e., non-block), the
power and expressiveness of the keyword system allow it to swap in block
semantics at render time.

In order to do this and remain in the good graces of the HTMLBars
semantics around idempotent re-renders, these keywords must precisely
implement the same semantics as HTMLBars itself. Namely, they must:

1. Determine if it is a first time render
2. Allow the keyword to set up state to determine stability
3. Determine whether the render node is stable (that is, is the
   top-level DOM created last time for this render node still valid).
   If the node itself is stable, it must continue revalidating any child
   nodes.
4. Determine whether the render node should be pruned (i.e. it had a
   previous representation, but it is now empty)
5. Finally, if the node was unstable or it is a first-time render,
   proceed to render the initial state and insert it into the DOM.

Note that user-land helpers are always a pluggable way to implement
logic that follows these rules, but they are very constrained. This
allows them to be implemented as "pure functions", without the user
having to worry about questions of stability or validation.

Keywords allow a host environment to provide abstractions like helpers,
where the user code can be considered to be executed over and over
again, but the implementation of the keyword itself is responsible for
maintaining DOM stability.

See the LIFECYCLE.md file for more information.

This commit makes it easier to implement keywords that satisfy the
idempotence requirement, without having to manually propagate child node
revalidation, pruning, and other considerations.